app/requirements.txt
# Backend
Django

# Infra
psycopg2-binary
gunicorn
sentry-sdk

# API
djangorestframework

# Static files
whitenoise

# Admin
django-extensions
ipython
